+++ test/acceptance/components/main-site-layout.js
@@ -3,7 +3,7 @@
 const articleData = require('../article-data.json');
 
 module.exports = {
-  name: 'MainSiteLayout',
+  type: 'MainSiteLayout',
   elements() {
     return [
       {
@@ -24,11 +24,11 @@ module.exports = {
   actions() {
     return {};
   },
-  children(instancename, options, expectedState) {
+  children(expectedState, dataStore) {
     return [
       {
-        componentName: 'NewsArticle',
-        instanceName: 'newsArticle1',
+        type: 'NewsArticle',
+        name: 'newsArticle1',
         state: {
           displayed: true,
           newsArticleImage: {
@@ -50,8 +50,8 @@ module.exports = {
         },
       },
       {
-        componentName: 'NewsArticle',
-        instanceName: 'newsArticle2',
+        type: 'NewsArticle',
+        name: 'newsArticle2',
         state: {
           displayed: true,
           newsArticleImage: {